refactor: 함수 분리
